TTS 코드 분리
