+++ apps/overlay/src/app/app.module.ts
@@ -1,11 +1,12 @@
 import { Module } from '@nestjs/common';
 import { PrismaModule } from '@project-lc/prisma-orm';
+import { OverlayModule } from '@project-lc/nest-modules';
 import { AppController } from './app.controller';
 import { AppGateway } from './app.gateway';
 import { AppService } from './app.service';
 
 @Module({
-  imports: [PrismaModule],
+  imports: [PrismaModule, OverlayModule],
   controllers: [AppController],
   providers: [AppService, AppGateway],
 })
