+++ libs/nest-modules/src/index.ts
@@ -9,3 +9,5 @@ export * from './lib/goods/goods.service';
 export * from './lib/_nest-units/middlewares/colorizedMorgan.middleware';
 export * from './lib/_nest-units/guards/jwt-auth.guard';
 export * from './lib/_nest-units/decorators/sellerInfo.decorator';
+export * from './lib/overlay/overlay.module';
+export * from './lib/overlay/overlay.service';
