+++ libs/nest-modules/src/lib/overlay/overlay.module.ts
@@ -1,0 +1,8 @@
+import { Module } from '@nestjs/common';
+import { OverlayService } from './overlay.service';
+
+@Module({
+  providers: [OverlayService],
+  exports: [OverlayService],
+})
+export class OverlayModule {}
